+++ project1/api/percents/__init__.py
@@ -1,0 +1,3 @@
+from .views import router
+
+__all__ = ["router"]
