+++ project1/api/percents/dtos.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class Percents(BaseModel):
+    value: float
+    percent: float
