+++ project1/api/users/__init__.py
@@ -1,0 +1,4 @@
+from .views import router
+
+
+__all__ = ["router"]
